refactor(home): fetch home data via route loader
