--- apps/frontend/portal/src/api/home.queryKeys.ts
@@ -1,4 +1,0 @@
-export const homeQueryKeys = {
-  all: ['home'] as const,
-  content: () => [...homeQueryKeys.all, 'content'] as const,
-};
