+++ apps/frontend/portal/src/features/Home/api/home.queries.ts
@@ -1,20 +1,11 @@
-import { useQuery } from '@tanstack/react-query';
+import { queryOptions } from '@tanstack/react-query';
 
-import { fetchRescueCases } from '../../RescueCases/api/rescueCases.api';
-
-import { getAdoptablePets } from './home.api';
+import { getHomeData } from './home.api';
 import { homeQueryKeys } from './home.queryKeys';
 
-export const useFetchLatestRescues = (limit: number) => {
-  return useQuery({
-    queryKey: homeQueryKeys.latestRescues(limit),
-    queryFn: () => fetchRescueCases(limit),
+export const homeQueryOptions = () =>
+  queryOptions({
+    queryKey: homeQueryKeys.content(),
+    queryFn: getHomeData,
+    staleTime: Infinity,
   });
-};
-
-export const useFetchAdoptablePets = () => {
-  return useQuery({
-    queryKey: homeQueryKeys.adoptablePets(),
-    queryFn: getAdoptablePets,
-  });
-};
