+++ apps/frontend/portal/src/features/Home/api/home.queryKeys.ts
@@ -1,6 +1,4 @@
 export const homeQueryKeys = {
   all: ['home'] as const,
-  latestRescues: (limit: number) =>
-    [...homeQueryKeys.all, 'latestRescues', limit] as const,
-  adoptablePets: () => [...homeQueryKeys.all, 'adoptablePets'] as const,
+  content: () => [...homeQueryKeys.all, 'content'] as const,
 };
