+++ apps/frontend/portal/src/features/Home/components/Hero.tsx
@@ -1,17 +1,15 @@
+import type { HomeData } from '@pawhaven/shared/types';
 import { ChevronRight } from 'lucide-react';
 import { preload } from 'react-dom';
 import { Trans, useTranslation } from 'react-i18next';
-import { useNavigate } from 'react-router-dom';
-
-import { EMPTY_HERO_STATS, useHomeData } from '@/api/home.queries';
+import { useLoaderData, useNavigate } from 'react-router-dom';
 
 preload('/images/hero-banner.webp', { as: 'image', type: 'image/webp' });
 
 export const Hero = () => {
   const { t, i18n } = useTranslation();
   const navigate = useNavigate();
-  const { data: homeData } = useHomeData();
-  const heroStats = homeData?.heroStats ?? EMPTY_HERO_STATS;
+  const { heroStats } = useLoaderData() as HomeData;
 
   const stats = [
     {
