+++ apps/frontend/portal/src/features/Home/index.tsx
@@ -1,4 +1,5 @@
-import { useNavigate } from 'react-router-dom';
+import type { HomeData } from '@pawhaven/shared/types';
+import { useLoaderData, useNavigate } from 'react-router-dom';
 
 import { RescueCasesSection } from '../RescueCases/components/RescueCasesSection';
 
@@ -6,13 +7,9 @@ import { AdoptablePetsSection } from './components/AdoptablePetsSection';
 import { Hero } from './components/Hero';
 import { StrayCTA } from './components/StrayCTA';
 
-import { useHomeData } from '@/api/home.queries';
-
 export const Home = () => {
   const navigate = useNavigate();
-  const { data: homeData } = useHomeData();
-  const latestRescues = homeData?.latestRescues ?? [];
-  const adoptablePets = homeData?.adoptablePets ?? [];
+  const { latestRescues, adoptablePets } = useLoaderData() as HomeData;
 
   const handleCaseClick = (id: string) => {
     navigate(`/rescue/detail/${id}`);
