+++ apps/frontend/portal/src/features/Home/route.tsx
@@ -1,6 +1,16 @@
+import { homeQueryOptions } from './api/home.queries';
+
 import { Home } from '@/features/Home';
+import { getQueryClient } from '@/providers/QueryProvider';
+
+export const homeLoader = async () => {
+  const queryClient = getQueryClient();
+
+  return queryClient.ensureQueryData(homeQueryOptions());
+};
 
 export const homeRoute = {
   index: true,
+  loader: homeLoader,
   Component: Home,
 };
